+++ src/main/java/com/pricewagon/pricewagon/domain/product/repository/ProductRepository.java
@@ -10,12 +10,12 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.pricewagon.pricewagon.domain.product.entity.Product;
 import com.pricewagon.pricewagon.domain.product.entity.type.ShopType;
 
-public interface ProductRepository extends JpaRepository<Product, Integer>, ProductRepositoryCustom {
+public interface ProductRepository extends JpaRepository<Product, Long>, ProductRepositoryCustom {
 	Page<Product> findAllByShopType(ShopType shopType, Pageable pageable);
 
 	Optional<Product> findByShopTypeAndProductNumber(ShopType shopType, Integer productNumber);
 
-	Page<Product> findByShopTypeAndCategory_IdIn(ShopType shopType, List<Integer> categoryIds, Pageable pageable);
+	Page<Product> findByShopTypeAndCategory_IdIn(ShopType shopType, List<Long> categoryIds, Pageable pageable);
 
 	boolean existsByProductNumber(Integer productNumber);
